requirements.txt
python-can
cantools
matplotlib
